Fix bar/scatter colour problems.
Set the bar marker config to a separate config object. When preparing a trace, make sure we just use trace items appropriate to the graph type.

This fixes config confusion where scatter config was leaking into bars meaning, for example, that one had to set the scatter colour to solid to get the bar drawing and avoid errors due to a lookup being done on the value for the ramp colour that is the scatter default.